Modify id of the Plan entity as Long type, start creating Plan screen

+++ app/src/main/java/com/example/mylifearranger/feature_planner/data/data_source/PlanDao.kt
@@ -30,7 +30,7 @@ interface PlanDao {
 //    fun getTasksForDate(date: String): Flow<List<Task>>
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertPlan(plan: Plan): Int
+    suspend fun insertPlan(plan: Plan): Long
 
     @Insert
     suspend fun insertPlanTask(planTask: PlanTask)
